requirements.txt
Flask>=2.0.0
flask-cors==3.0.10
psycopg2-binary==2.9.9